+++ MS5837.h
@@ -23,9 +23,10 @@
 
 
 //  ERROR CODES
-#define MS5837_OK                 0x00
-//  TODO elaborate
-//  I2C can return 2, 3, 4
+#define MS5837_OK                   0
+//  I2C twoWire can return 2, 3, 4
+#define MS5837_ERROR_I2C          -10
+#define MS5837_ERROR_REQUEST      -11
 
 
 class MS5837
@@ -87,6 +88,15 @@ public:
   int      getLastError();
 
 
+  //////////////////////////////////////////////////////////////////////
+  //
+  //  PROM zero - meta info
+  //
+  uint16_t getCRC();
+  uint16_t getProduct();
+  uint16_t getFactorySettings();
+
+
 protected:
   int        command(uint8_t cmd);
   void       initConstants(uint8_t mathMode);
@@ -101,11 +111,10 @@ protected:
 
   float     C[8];
   uint8_t   _type = MS5837_TYPE_UNKNOWN;
-  int       _result;
 
   float     _density = 0.99802;  //  water at 20 °C
   //  prepare error handling.
-  int       _error = 0;
+  int       _error = MS5837_OK;
   uint32_t  _lastRead;
 };